Fix backup method not resetting board state
I added a visual debugger that made it readily apparent that my backup method
wasn't properly updating the visited state of the board, or the possible move
states. Simple fix, and now it seems to work exactly as intended and is fairly
efficient. Commiting the visual debugger in a separate commit as it can still
use some cleaning up and optimization.